0.1.9 -> 0.2.0
Added Custom Weapon Cycles. This is a much better solution to the problem that the mod is supposed to address. It is able to replicate all of the same behavior as before, and more, and is more maintainable. 
It still has some issues that will be fixed, so the old method of input binds still exist.